% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotModule.R
\name{BoxplotServer}
\alias{BoxplotServer}
\title{BoxplotServer}
\usage{
BoxplotServer(id, project, DF, Pheno, theme)
}
\arguments{
\item{id}{id}

\item{project}{project}

\item{DF}{DF}

\item{input}{input}

\item{pheno}{description}
}
\value{
nothing
}
\description{
BoxplotServer
}
\examples{
\dontrun{"non"}


}
