% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotModule.R
\name{BoxplotUI}
\alias{BoxplotUI}
\title{BoxplotUI user interface part for Loading data widget}
\usage{
BoxplotUI(id, Pheno)
}
\arguments{
\item{id}{ID to use in the UI}

\item{pheno}{pheno}
}
\value{
a module of loading data
}
\description{
BoxplotUI user interface part for Loading data widget
}
\examples{
\dontrun{"non"}

}
