% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ediTable.R
\name{ediTable}
\alias{ediTable}
\title{ediTable}
\usage{
ediTable(id, ...)
}
\arguments{
\item{id}{id in shiny app}

\item{...}{other rHandsontableOutput parametters (try ?rHandsontableOutput)}
}
\value{
ui output
}
\description{
ediTable
}
\examples{
"non"
}
