% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ediTable.R
\name{ediTable_server}
\alias{ediTable_server}
\title{ediTable_server}
\usage{
ediTable_server(
  id,
  rd,
  allowRowEdit = TRUE,
  allowColumnEdit = T,
  manualRowMove = TRUE,
  ...
)
}
\arguments{
\item{id}{id in shiny app}

\item{rd}{endred tble}

\item{allowRowEdit}{default T}

\item{allowColumnEdit}{default F}

\item{manualRowMove}{default T}

\item{...}{other rhandsontable parametters (try ?rhandsontable)}
}
\value{
server outputs
}
\description{
ediTable_server
}
\examples{
"non"
}
