% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMModule.R
\name{FC_HeatmapServer}
\alias{FC_HeatmapServer}
\title{BoxplotServer}
\usage{
FC_HeatmapServer(id, project, DF, Pheno, FC.geneCol)
}
\arguments{
\item{id}{id}

\item{project}{project}

\item{DF}{DF}

\item{FC.geneCol}{name of colnames correspondant au geneID du DF}

\item{pheno}{description}
}
\value{
nothing
}
\description{
BoxplotServer
}
\examples{
\dontrun{"non"}


}
