% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAHMModule.R
\name{GSEA_HeatmapServer}
\alias{GSEA_HeatmapServer}
\title{BoxplotServer}
\usage{
GSEA_HeatmapServer(id, project, DF, Pheno)
}
\arguments{
\item{id}{id}

\item{project}{project}

\item{DF}{DF}

\item{Pheno}{description}
}
\value{
nothing
}
\description{
BoxplotServer
}
\examples{
\dontrun{"non"}


}
