% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAHMModule.R
\name{GSEA_HeatmapUI}
\alias{GSEA_HeatmapUI}
\title{BoxplotUI user interface part for Loading data widget}
\usage{
GSEA_HeatmapUI(id, Height = 10, Width = 10)
}
\arguments{
\item{id}{ID to use in the UI}

\item{Height}{default 10cm}

\item{Width}{default 10cm}
}
\value{
a module of loading data
}
\description{
BoxplotUI user interface part for Loading data widget
}
\examples{
\dontrun{"non"}

}
