% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingData.R
\name{LoadingData}
\alias{LoadingData}
\title{LoadingData server part}
\usage{
LoadingData(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
nothing
}
\description{
LoadingData server part
}
\examples{
\dontrun{"non"}

}
