% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingData.R
\name{LoadingDataUI}
\alias{LoadingDataUI}
\title{LoadingDataUI user interface part for Loading data widget}
\usage{
LoadingDataUI(id, column.size = 12, label)
}
\arguments{
\item{id}{ID to use in the UI}

\item{column.size}{column size}

\item{label}{description}
}
\value{
a module of loading data
}
\description{
LoadingDataUI user interface part for Loading data widget
}
\examples{
\dontrun{"non"}

}
