% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFiles.R
\name{SamplesFileUI}
\alias{SamplesFileUI}
\title{Chose Sampels files UI user interface part for Loading data widget}
\usage{
SamplesFileUI(id, column.size = 12, label)
}
\arguments{
\item{id}{ID to use in the UI}

\item{column.size}{column size}

\item{label}{description}
}
\value{
a module of loading data
}
\description{
Chose Sampels files UI user interface part for Loading data widget
}
\examples{
\dontrun{"non"}

}
